Game exhibiting new functs of C++ and while true
This is a fun game that is called Race of Drunk Letters. In this game any letter from a to z of the user's choice races from the start to the finish like a drunk, sometimes going ahead and sometimes going backwards.
You will be introduced to using functions like: srand(), rand(), time(0). //related to generating random and pseudo-random sequences. And you will also see the use of while(true). All these concepts are used in raspberry-py and almost every online hardware as something called an "INTERRUPT", which I will go into in future programs.